[Feat] FavorTabBarController, 화면 중앙 선물 추가 하기 버튼으로 GiftManagement로 이동
