+++ Favor/Favor/Sources/Flows/AppFlow.swift
@@ -46,7 +46,7 @@ public final class AppFlow: Flow {
 
     case .authIsRequired:
       return self.navigateToAuth()
-
+      
     case .localAuthIsRequired(let location):
       return self.navigateToLocalAuth(request: location)
 
@@ -55,7 +55,10 @@ public final class AppFlow: Flow {
 
     case .wayBackToRootIsRequired:
       return self.navigateWayBackToRoot()
-
+      
+    case .giftManagementIsRequired:
+      return self.navigateToGiftManagement()
+      
     default:
       return .none
     }
@@ -166,6 +169,22 @@ private extension AppFlow {
   func navigateWayBackToRoot() -> FlowContributors {
     return .one(flowContributor: .forwardToCurrentFlow(withStep: AppStep.authIsRequired))
   }
+  
+  func navigateToGiftManagement() -> FlowContributors {
+    let newGiftFlow = NewGiftFlow()
+
+    Flows.use(newGiftFlow, when: .ready) { [unowned self] root in
+      DispatchQueue.main.async {
+        root.modalPresentationStyle = .overFullScreen
+        self.rootViewController.present(root, animated: true)
+      }
+    }
+
+    return .one(flowContributor: .contribute(
+      withNextPresentable: newGiftFlow,
+      withNextStepper: OneStepper(withSingleStep: AppStep.giftManagementIsRequired())
+    ))
+  }
 }
 
 // MARK: - Contributors
@@ -192,7 +211,8 @@ private extension AppFlow {
       .contribute(
         withNextPresentable: myPageFlow,
         withNextStepper: OneStepper(withSingleStep: AppStep.myPageIsRequired)
-      )
+      ),
+      .contribute(withNext: self.rootViewController)
     ]
   }
 }
